requirements.txt
# FastAPI & Web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Celery & Message Queue
celery==5.3.6
redis==5.0.1
kombu==5.3.4

# Socket.IO
python-socketio==5.10.0
aioredis==2.0.1

# ML/Deep Learning (chỉ cần cho worker)
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
SimpleITK==2.3.1
matplotlib==3.7.2
scikit-image==0.21.0
scipy==1.11.3

# PyTorch (cài riêng với CUDA)
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# Transformers
# diffusers==0.21.4
# transformers==4.35.2
# accelerate==0.24.1

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1